+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### dependabot: \#31 Bump ip from 1.1.8 to 1.1.9
+
 ## [0.4.0] - 2024-01-29
 
 - added `CacheProvider<TEnum>` utility class